# openapi.yaml
openapi: 3.0.3
info:
  title: StudySync API
  description: Backend API for the StudySync collaborative learning platform. Reliability score endpoints are not included.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        dueDate:
          type: string
          format: date
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date
        variant:
          type: string
          enum: [default, highlight, completed]
        createdBy:
          type: object
          properties:
            id:
              type: string
            initials:
              type: string
            name:
              type: string
            color:
              type: string
        pendingRegressRequest:
          type: object
          properties:
            requestId:
              type: string
            requesterId:
              type: string
            fromStatus:
              type: string
            targetStatus:
              type: string
            createdAt:
              type: string
              format: date-time
        assignee:
          type: object
          properties:
            id:
              type: string
            initials:
              type: string
            name:
              type: string
            color:
              type: string
        createdAt:
          type: string
          format: date-time
    KanbanBoard:
      type: object
      properties:
        todo:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        in_progress:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        completed:
          type: array
          items:
            $ref: '#/components/schemas/Task'
...
paths:
...
  /workspaces/{groupId}:
    get:
      tags: [Workspaces]
      summary: Workspace metadata and members
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
  /workspaces/{groupId}/tasks:
    get:
      tags: [Workspaces]
      summary: Kanban board
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KanbanBoard'
    post:
      tags: [Workspaces]
      summary: Create task
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                dueDate:
                  type: string
                assigneeId:
                  type: string
      responses:
        '201':
          description: Created
  /workspaces/{groupId}/tasks/{taskId}:
    patch:
      tags: [Workspaces]
      summary: Update or move task
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated
    delete:
      tags: [Workspaces]
      summary: Delete task
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '403':
          description: Only the task creator can delete
  /workspaces/{groupId}/tasks/reorder:
    put:
      tags: [Workspaces]
      summary: Reorder or move tasks (backward moves by non-creators return 409 REGRESS_REQUIRES_APPROVAL)
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated board
  /workspaces/{groupId}/tasks/{taskId}/progress:
    post:
      tags: [Workspaces]
      summary: Assignee marks task started or complete (auto-moves column)
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [start, complete]
      responses:
        '200':
          description: Updated task
  /workspaces/{groupId}/tasks/{taskId}/regress-requests:
    post:
      tags: [Workspaces]
      summary: Request creator approval to move a task to an earlier column
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetStatus]
              properties:
                targetStatus:
                  type: string
                  enum: [todo, in_progress]
      responses:
        '201':
          description: Request created; creator notified
  /workspaces/{groupId}/tasks/{taskId}/regress-requests/{requestId}/approve:
    post:
      tags: [Workspaces]
      summary: Approve a regress request (creator only)
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: taskId
          in: path
          required: true
          schema:
            type: string
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Approved; all pod members notified
  /workspaces/{groupId}/tasks/{taskId}/regress-requests/{requestId}/reject:
    post:
      tags: [Workspaces]
      summary: Reject a regress request (creator only)
      security:
        - bearerAuth: []
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
        - name: taskId
          in: path
          required: true
          schema:
            type: string
        - name: requestId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Rejected; requester notified
...